+++ src/rdmnet/discovery/monitored_scope.h
@@ -20,8 +20,15 @@
 #ifndef MONITORED_SCOPE_H_
 #define MONITORED_SCOPE_H_
 
-typedef struct RdmnetScopeMonitorRef RdmnetScopeMonitorRef;
+#include "rdmnet/core/discovery.h"
+#include "discovered_broker.h"
+#include "disc_platform_defs.h"
 
+#ifdef __cplusplus
+extern "C" {
+#endif
+
+typedef struct RdmnetScopeMonitorRef RdmnetScopeMonitorRef;
 struct RdmnetScopeMonitorRef
 {
   // The configuration data that the user provided.
@@ -36,5 +43,17 @@ struct RdmnetScopeMonitorRef
   // The next ref in the list of scopes being monitored.
   RdmnetScopeMonitorRef* next;
 };
+
+RdmnetScopeMonitorRef* scope_monitor_new(const RdmnetScopeMonitorConfig* config);
+void scope_monitor_insert(RdmnetScopeMonitorRef* scope_ref);
+bool scope_monitor_ref_is_valid(const RdmnetScopeMonitorRef* ref);
+void scope_monitor_for_each(void (*for_each_func)(RdmnetScopeMonitorRef*));
+void scope_monitor_remove(const RdmnetScopeMonitorRef* ref);
+void scope_monitor_delete(RdmnetScopeMonitorRef* ref);
+void scope_monitor_delete_all();
+
+#ifdef __cplusplus
+}
+#endif
 
 #endif /* MONITORED_SCOPE_H_ */
