+++ tests/unit/discovery/common/disc_platform_defs.h
